fix: redirect only happening on first load on the root path
