+++ frontend/src/component/InitialRedirect.tsx
@@ -1,0 +1,42 @@
+import { useCallback, useEffect, useContext } from 'react';
+import { useNavigate } from 'react-router-dom';
+import useProjects from '../hooks/api/getters/useProjects/useProjects';
+import { useLastViewedProject } from '../hooks/useLastViewedProject';
+import UIContext from 'contexts/UIContext';
+import Loader from './common/Loader/Loader';
+
+export const InitialRedirect = () => {
+    const { lastViewed } = useLastViewedProject();
+    const { projects, loading } = useProjects();
+    const { redirected, setRedirected } = useContext(UIContext);
+    const navigate = useNavigate();
+
+    // Redirect based on project and last viewed
+    const getRedirect = useCallback(() => {
+        if (projects && lastViewed) {
+            return `/projects/${lastViewed}`;
+        }
+
+        if (projects && !lastViewed && projects.length === 1) {
+            return `/projects/${projects[0].id}`;
+        }
+
+        return '/projects';
+    }, [lastViewed, projects]);
+
+    const redirect = () => {
+        setRedirected(true);
+        navigate(getRedirect(), { replace: true });
+    };
+
+    useEffect(() => {
+        if (redirected) return;
+        redirect();
+    }, [getRedirect]);
+
+    if (loading) {
+        return <Loader />;
+    }
+
+    return null;
+};
