+++ frontend/src/component/providers/UIProvider/UIProvider.tsx
@@ -30,6 +30,7 @@ const UIProvider: React.FC<IUiProviderProps> = ({
 }) => {
     const [toastData, setToast] = useState<IToast>(createEmptyToast());
     const [showFeedback, setShowFeedback] = useState(false);
+    const [redirected, setRedirected] = useState(false);
     const [themeMode, setThemeMode] = useState(resolveMode(darkmode));
 
     const context = React.useMemo(
@@ -40,8 +41,10 @@ const UIProvider: React.FC<IUiProviderProps> = ({
             setShowFeedback,
             themeMode,
             setThemeMode,
+            redirected,
+            setRedirected,
         }),
-        [toastData, showFeedback, themeMode]
+        [toastData, showFeedback, themeMode, redirected]
     );
 
     return <UIContext.Provider value={context}>{children}</UIContext.Provider>;
