+++ frontend/src/component/user/DemoAuth/DemoAuth.tsx
@@ -28,7 +28,7 @@ const DemoAuth: VFC<IDemoAuthProps> = ({ authDetails, redirect }) => {
         try {
             await emailAuth(authDetails.path, email);
             refetchUser();
-            navigate(redirect);
+            navigate(redirect, { replace: true });
         } catch (error) {
             setToastApiError(formatUnknownError(error));
         }
